refactor: WorkerMessengerFactory
